add newline at EOF - POSIX compliance
POSIX [1] defines a line as follows:

---
3.206 Line

A sequence of zero or more non- <newline> characters plus a terminating <newline> character.
---

Therefore, lines not ending in a newline character aren't considered actual lines.
That's why some programs have problems processing the last line of a file if it isn't newline terminated.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_206