Desktop: Fixes getting latest version for Linux Installer
For some reason I was getting an error on evaluating the expression
that was fetching latest version of joplin in the Linux installer
script.

I'm using zsh so this could be the case. Thought I haven't debugged
this much so it could be related to something else.

Changing the last grep part fixed the issue and I could install
joplin without errors.