add option to show line numbers in editor
This also allows code folding in the editor.

The editor options `showLineNumbers` and `showFoldWidgets` are true by default.
But they only show up, if the gutter is displayed, which is explicitly disabled in Joplin.
Therefore this change onlu toggles the gutter.

see https://discourse.joplinapp.org/t/feature-request-folding-sections-of-text/5752?u=tessus