Allow context menu to display options for both copy text and copy URL
Issue #3404 points out that on MacOS, right clicking on a link
only displays the 'Copy' context menu item, instead of 'Copy Link'.
It appears that MacOS selects the text while secondary clicking so
the currently implemented conditional exits before realizing a link
is also present. This change will allow context menu items for both
copying text and the link URL. It would likely be cleaner to have
ContextMenuItemType enum values in an array, and display options
based on the items in the array, but I'm worried about unintended
consequences of that big of a refactor since I don't really see tests
around any of this.