Update terminal.md
There is no objective reason to keep this if it's not supported.

It would be awesome to have a html directory or file automatically regenerated every time I change my joplin files. I was writing a script for this in the assumptions that html-export would be possible only to get the disappointing message "HTML export is not supported. Please use the desktop application.".

Why have it in the docs for the cli app? We might as well add jpeg, wav and psd and then add more notifications that these export methods are not supported.