Link to experimental OneNote export app
Hi, not sure if this is the right way to propose it: I wrote a OneNote export tool for the web browser for my migration to Joplin. So far only tested with my account (~100MB), so it is experimental. Maybe useful for others? It is open source and data stays completely in the user's browser.