update default detached HEAD branch (master -> dev)
We are no longer building releases from master.
Travis CI runs from a detached HEAD state, which makes it impossible to tell the branch.

I still think that it is imoprtant to show the branch in the About window.
It helps narrowing down issues should people compile from their own branch and it helps developers to see on which branch they are running.

This change only changes the name of the branch to `dev` in a detached state.