Add a top-level "watch" config.
It configures the "build_clipper" task to only run when the following files and directories change: content_scripts, popup/build, icons, background.js, manifest.json. This "build_clipper" task is also delayed for 6000 milliseconds before checking for file changes. This is to ensure that it will be run after the script that builds popup has finished.