Introduce a script that builds an optimized folder for installing unpacked extension.
This script copies only the necessary files and folders in app-clipper to a subdirectory at app-clipper/build/{browserName}. The files and folders to be copied are: background.js, content_scripts/, icons, manifest.json, popup/build.

The script also removes the key in manifest.json that will cause Chrome or Firefox to throw an error.