Configured User-Agent header and handle throttling errors
User-Agent is pulled from a packageInfo.js file - this file is created dynamically through a gulp task at build time, similar to the packageInfo.js file in /packages/app--desktop. This is an attempt to reduce throttling based on Microsoft documentation: https://docs.microsoft.com/en-us/sharepoint/dev/general-development/how-to-avoid-getting-throttled-or-blocked-in-sharepoint-online#how-to-decorate-your-http-traffic-to-avoid-throttling

Updates to the error handling on exec() requests when the onedrive throttle code 'activityLimitReached' is present. In this instance, the retry-after HTTP header is used to pause the sync thread and the request retry count is decremented as this is an error that is expected to resolve itself in time.