Add test for newlines in title
Remove whitespace stripping
The original reason for this is that non-technical users aren't used
to whitespace sensitivity so it made sense to ignore whitespace.
This, however, breaks multiline yaml strings.
I realized that the original reasoning is flawed, this feature needs
to be accessible to non-technical users, but editing/creating the yaml
will still be in the domain of technical users.