Set MAILER_SECURE default to false to align with nodemailer default
Any value other an an explicit "1" is interpreted as false.

Fixes #5222

MAILER_SECURE enables a forced SSL mode (SMTPS) often seen on port 465.
The other common variant of port 587 & STARTSSL requires MAILER_SECURE
to be false. More importantly, as previously written there was no way
to set it to false, thus making STARTTLS impossible to use. With this
setup both options can be used as required in the installation
environment.