Desktop: fix wording in settings and also allow to check for updates on Linux
Currently the wording in settings leas people to believe that the applications will be automatically updated.
In reality all that is done is checking for available updates.
For some reason this check was disabled on Linux, because during experiments with real auto-updates yielded negative results on Linux.
But since the current code only checks for updates, there is no reason to not allow this on Linux.