Clarify where to go to enable markdown plugins
A new user on his first day might lose 3 hours trying to enable, say, the Fountain plugin, after seeing it's actually *not* in the plugins list. What is called "plugins" here, are actually *markdown plugins*, completely different from the main type of plugins. To enable a markdown plugin, user should not search for it in the "plugins" section of the options, but go to the "markdown" section of the options, and just tick the right box.