Fix-5968: Default sort order lost on exit
This fixes the bug #5968; When a notebook with its own sort order is selected and the app exits, the default (shared) sort order is lost at the next startup of the app.

The cause of the bug is that the initial notebook is not tracked. So, at the first time that notebook is switched, the cleanup of the last notebook is not performed.