Use apt to install tini to enable multi-platform support
The way tini was being installed broke multi-platorm (amd64, armv7, arm64) support.

```bash
# Download the init tool Tini and make it executable for use in the final image
ADD https://github.com/krallin/tini/releases/download/v0.19.0/tini-static /tini
```

This is a hard reference to the amd64 version of tini, which isn't compatible with armv7 and arm64. Tini has builds for other platforms though... The easiest way to fix this is to simply use the apt package manager to install tini as that will result in retrieval of the correct binary.