Spell Check: Optimize check if we can set the language with certain code
The problem was originally from issue #4146. The cause of the issue is
that on macOS Electron function `getSpellCheckerLanguages` could return
language code, that is not supported by `setSpellCheckerLanguages`. Also
based on Electron documentation this two functions is no-op on macOS,
since it uses built-in spell checker, but they clearly not no-op, based
on the issues found on github like this one
https://github.com/electron/electron/issues/30215. During the initial
setup, if user first runs the app, and doesn't have
spellChecker.language set, we use `getSpellCheckerLanguages` to get
system language and set spell check to it. But since macOS could return
not supported code, the error will occur when trying to set it.

The workaround was to try to set the language, and if it trow a error,
try to set nearest supported language. This actually isn't needed, since
the we can check supported languages from
`availableSpellCheckerLanguages`, and it will always return languages,
that supported by `setSpellCheckerLanguages`. Since they are hardcoded
in the Chromium source code.
https://source.chromium.org/chromium/chromium/src/+/main:components/spellcheck/common/spellcheck_common.cc;l=30;drc=f1a93e5d28bef5c70d5552758d7aa6b85bc54418;bpv=0;bpt=1

So we can just check if the result of `getSpellCheckerLanguages` in the
list of `availableSpellCheckerLanguages`, and if not set it to our
default (en-US).