Refactor code,make menu width progressively larger
The logic for getting the side menu width was being
repeated in two parts of the code, so I created a
function for that. Also, I made the side menu width
increase with each successive breakpoint to take
advantage of the screen width to display longer
note titles without truncating it.