Fix notesbar update issues, Simply notesbar search
Previously the notesbar wasn't being updated, when the note changed from inside the note editor, this commit fixes that. Also, the search functionality worked similarly to that of the search screen, this
commit changes it to a very simple search using the note titles, the
search will be improved on in the future.