modified the existing docker-compose.server-dev.yml file to add modifications:
Add a `restart` policy to the `app` and `db` services. This will ensure that the containers are restarted if they stop or crash for any reason.

Add a `depends_on` directive to the app service. This will ensure that the db service is started before the app service, as the app service depends on the database to function.

Add a `volumes` section to the db service. This will allow you to persist the data in the database even if the container is stopped or deleted.

Add a `command` to the app service. This will allow you to specify the command that is run when the container is started.

Add a `healthcheck` to the app service. This will allow you to specify a command that is used to check the health of the service, and if it fails, the container will be restarted.