Desktop: Make note sort update logic use correct prior sort and drop-grouping
When notes are being dragged & dropped, the new custom order values are calculated
on the basis of the new surrounding note IDs; to determine the new surrounding note
IDs, the notes are re-sorted. This sorting operation must use the same sorting
and filtering rules as the note list display, for the insert to work correctly.

Furthermore, when uncompleted notes are positioned at the start, the "sort step"
logic must be adjusted to treat completed/non-todo notes and uncompleted ones as
entirely separate windows.