renderer/tests: Test createEventHandlingAttrs and migrate to ts-jest
To test some aspects createEventHandlingAttrs, jsdom was required. The current
transpilation from TypeScript to JavaScript adds a "use strict" to the
top of transpiled files, however, a comment declaring that a file's
environment should be js-dom must be at the top of a file to take
effect. Thus, transpiling tests moves this comment away from the top of
the file, causing these tests to not use jsdom.