Fix error handeling in server response JSON parsing
In case the server is not responding with a JSON file type we try to
parse the plain text as JSON. If the plain text is not JSON (likely
HTML), a generic and hard-to-debug error was generated.
This commit adds to the error log the first ~1000 characters of the
plain text using the project's error format instead of the basic JS
exception error.