Update README.md
Request to explicitly define flags that are supported by the Joplin_install_and_update.sh script. I did not realize that there was an option to update to the prerelease version until seeing discord post https://discourse.joplinapp.org/t/joplin-cloud-not-syncing/30844/6. One could always review the contents of the script, however it would be great to be informed of the supported flag options during initial install on Linux.