Revert "Debugging: Attempt to run android build workflow on MacOS"
This reverts commit 21959a1ccd1721d7e254110d581c3aff7ba32876.

The MacOS workflow failed due to an outdated version of Java (the
workflow requires Java 11, but the CI Mac would need to install this).
For now, I'm focusing on setting up the Android build on my local Mac (I
suspect getting this to work will be faster than getting the CI to
work).