Desktop: Fixes #8409: Only attempt to decrypt one time when clicking "retry"
Prior to this commit, clicking "retry" on an item that failed to decrypt
in the sync details screen removed the item from the list, regardless of
whether the decryption retry failed.

This commit disables decryption for a resource if retrying decryption
fails, thus adding it back to the disabled list.