Move appClose handler to app.ts
With the appClose event handler in `MainScreen.tsx`, changing screens
(e.g. opening settings) caused the `MainScreen` component to be replaced
by another.

In this case, after returning from settings, React constructs a second
copy of `MainScreen`, which registers the event handler again. Because
we want this event handler to be triggered even if the settings screen
is open, it has been moved to a location where it won't be re-loaded on
screen change — to `app.ts`.