Turndown package: Update Rollup
The old version of rollup used by turndown seems to not recognize the
`\P` unicode escape. This commit updates Rollup to the latest version.

Inspection of the generated output reveals `require`s in each file. This
seems to have also been the behavior before this commit, and is likely
caused by the usage of `require` instead of `import` in some of the
`turndown` source files.