Added option to show full id in ls command
As the ls command only shjows either the title, or the title, short id
and date, there is no way to programatically obtain the necessary id to
create a joplin internal URL in scripts.

Added the -i --id option to the ls command, which, by itself, shows the
full id in parenthesis after the title, or if combined with the -l
--long option, it just shows the full id, so scripts can extract them
and create valid internal links (:/[id]).