FIX laurent22#9970: Focus is lost when the hyperlink modal is cancelled
When a hyperlink is entered and confirmed with <Enter> or by clicking OK, everything works as expected. However, when canceling with <Esc> or the Cancel button,
the focus shifts away from the pane. While Joplin seems to remain in focus, as evidenced by the ability to access the File menu by pressing <Alt>, it requires
clicking back into the pane to resume typing. Alternatively, pressing <Shift-Tab> three times or <Tab> four times refocuses the Markdown pane.
I identified the issue in the prompt function of the Smalltalk library. To address this, I created a wrapper called SmalltalkFocusWrapper to resolve the focus
loss problem upon canceling the modal. Now, when canceling the modal, the focus remains in the appropriate location.