FIX laurent22#9950: Link created by LINKIFY in WYSIWYG editor is not underlined until switch to another note
The current behavior is such that when adding a link to a note using the WYSIWYG editor, the link is not immediately underlined.
However, if you switch to another note and then return to the note created earlier, the link is underlined as expected.
The expected behavior is that links should be underlined immediately upon insertion, without the need to leave and return to the note.

I found that the problem was that when I pasted the link it was considered as plainText, so I created a function that checks if the pasted
text is a link and added this check to the markdowntags if.