Desktop: Fixes #9970: focus is lost when the hyperlink modal is cancelled
When closing the hyperlink prompt dialog, either by clicking cancel or just
by pressing escape, the prompt function returns a promise that is never
resolved or rejected, hence not executing the following code which would
give back focus to the editor.

This behaviour happens because the smalltalk library checks for a cancel
option. If this option is false or undefined (which was always the case),
the promise that is returned is never cancelled/rejected. What I did was
set this option to true so that the promise can be rejected and therefore
continue the flow of execution.

Signed-off-by: Fabio Neto <fabiogvdneto@tecnico.ulisboa.pt>