Desktop: Fixes #8960: Fix cursor jumps to the top of the note on sync
This commit fixes a race condition caused by the editor's form note's hasChanged
property not being up-to-date (or checked) while refreshing the note.

The `.hasChanged` property was previously checked only before running
async logic that updated the note. If the note changed during this async
logic, this could change the value passed down to the editor as its
content. If the editor had recently updated its content, the content
prop would then **not match** the current content of the editor.

The CodeMirror 5 (and also the rich text?) editors handle this by 1)
changing the editor's content 2) clearing the editor's undo history and
3) resetting the cursor position. The CodeMirror 6 editor only does 1
and 2, so the issue was most significant in the CodeMirror 5-based
editor.

This commit fixes the issue by 1) storing a refrence to the latest
formNote in a ref and 2) checking ref.current.hasChanged just before
saving changes to the note.