Desktop: Show notification on import/export using interop service
When importing/exporting items in the desktop app, a giant
model layer covers all our screen, and the users are forced
to wait until the process is completed. This raised our
attention and the need of implementing something that would let
the users do their things, while still being able to know about
when the process is happening and when it is finished.

We though and decided that one possible solution would be using
the notyf library, which right now is only used by the
TrashNotification component, and show a simple and minimal
notification indicating the state of the process, instead of
covering the entire screen with a modal layer. We truly
believe that this would drastically increase the user
experience.

Co-authored-by: Henrique Silva <henriqueapsilva@tecnico.ulisboa.pt>