Desktop: Fixes #10416: Work around checkbox won't change when editor is readonly
Works around an upstream change in the codemirror-vim package (which we
use for CM5 emulation). This **is a workaround**.

Currently, we use `codemirror-vim` to provided a polyfill for several
CodeMirror 5 methods, including `replaceSelection`. `codemirror-vim` now
disables some of these when the editor is read-only. (Ideally, we should
not depend on this compatibility layer for core application logic).