Show "ctrl+click to open" tooltip over links
This is a potential fix for #10199. This particular fix displays a
custom tooltip, in addition to the OS tooltip. The custom tooltip is
shown above the target element, while the OS tooltip (at least on my
Linux system) is shown below.

After this, I plan to work on a different fix that changes the content
of the OS tooltip, and changes it back before saving. This would be a
similar approach to #10359, except change the content in the rich text
editor, rather than in `turndown`. This has the advantage of also
working for HTML notes.