Fix #9588: Rich text editor: Mermaid download button doesn't work
Added an additional verification that will only allow the mermaid
diagram download button to appear if the user is using the markdown
editor. It uses the document object, which represents the
loaded html page, and the function querySelector to look for the
first instance of use of the tox-tinymce class, exclusively used in
the rich text editor. If the return value of this verification is
not null, the constant exportButtonMarkup will simply be an empty
string, instead of calling the exportGraphButton function. This
way, the download button will not be visible in the rich text
editor.